requirements.txt
Flask==3.0.0
kafka-python==2.0.2
pyspark==3.5.0
websocket-client==1.6.4
numpy==1.26.2
pandas==2.1.3
requests==2.31.0